transaction-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
httpx==0.25.2
pytest==7.4.3
prometheus-fastapi-instrumentator==6.1.0
